// types/EventsResponse.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface EventsResponse {
  asset_events: AssetEvent[];
}

export interface AssetEvent {
  approved_account: null;
  asset: Asset;
  asset_bundle: null;
  auction_type: null;
  bid_amount: null;
  collection_slug: string;
  contract_address: string;
  created_date: string;
  custom_event_name: null;
  dev_fee_payment_event: null;
  duration: null;
  ending_price: null;
  event_type: string;
  from_account: FromAccount;
  id: number;
  is_private: null;
  owner_account: null;
  payment_token: null;
  quantity: string;
  seller: null;
  starting_price: null;
  to_account: FromAccount;
  total_price: null;
  transaction: null;
  winner_account: null;
}

export interface Asset {
  id: number;
  token_id: string;
  num_sales: number;
  background_color: null;
  image_url: string;
  image_preview_url: string;
  image_thumbnail_url: string;
  image_original_url: null;
  animation_url: null;
  animation_original_url: null;
  name: string;
  description: null;
  external_link: null;
  asset_contract: AssetContract;
  permalink: string;
  collection: Collection;
  decimals: null;
  token_metadata: null;
  owner: FromAccount;
}

export interface AssetContract {
  address: string;
  asset_contract_type: string;
  created_date: string;
  name: string;
  nft_version: null;
  opensea_version: string;
  owner: number;
  schema_name: string;
  symbol: string;
  total_supply: null;
  description: null;
  external_link: null;
  image_url: null;
  default_to_fiat: boolean;
  dev_buyer_fee_basis_points: number;
  dev_seller_fee_basis_points: number;
  only_proxied_transfers: boolean;
  opensea_buyer_fee_basis_points: number;
  opensea_seller_fee_basis_points: number;
  buyer_fee_basis_points: number;
  seller_fee_basis_points: number;
  payout_address: null;
}

export interface Collection {
  banner_image_url: null;
  chat_url: null;
  created_date: string;
  default_to_fiat: boolean;
  description: string;
  dev_buyer_fee_basis_points: string;
  dev_seller_fee_basis_points: string;
  discord_url: null;
  display_data: DisplayData;
  external_url: null;
  featured: boolean;
  featured_image_url: null;
  hidden: boolean;
  safelist_request_status: string;
  image_url: string;
  is_subject_to_whitelist: boolean;
  large_image_url: null;
  medium_username: null;
  name: string;
  only_proxied_transfers: boolean;
  opensea_buyer_fee_basis_points: string;
  opensea_seller_fee_basis_points: string;
  payout_address: string;
  require_email: boolean;
  short_description: null;
  slug: string;
  telegram_url: null;
  twitter_username: null;
  instagram_username: null;
  wiki_url: null;
}

export interface DisplayData {
  card_display_style: string;
}

export interface FromAccount {
  user: User;
  profile_img_url: string;
  address: string;
  config: string;
}

export interface User {
  username: null | string;
}